# api/src/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Bon-API
  version: v0
servers:
  -
    url: 'http://localhost:8000/api/v0'
paths:
...
  /tableReservation:
    get:
      tags:
        - 'Table Reservation'
      operationId: c9ae1535f885a3236309511473dc5e8d
      responses:
        '200':
          description: 'Retourne l\'' ensemble des tables réservées'
          content:
            application/json:
              schema:
                description: 'Toutes les tables réservées'
                type: array
                items:
                  $ref: '#/components/schemas/TableReservation'
    post:
      tags:
        - 'Table Reservation'
      operationId: 14382974d48ba13fc75b1fab0e26ae49
      requestBody:
        content:
          application/json:
            schema:
              required:
                - table_id
                - reservation_id
              properties:
                table_id:
                  type: intgeger
                reservation_id:
                  type: integer
              type: object
              example:
                table_id: 3
                reservation_id: 2
      responses:
        '200':
          description: 'Table réservée créée'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
  '/tableReservation/{id}':
    get:
      tags:
        - 'Table Reservation'
      operationId: f67fd71f80411450b9ee00026611647f
      parameters:
        -
          name: id
          in: path
          description: 'ID du table liée à une reservation'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Retourne une table liée à une réservation en fonction de son ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableReservation'
    put:
      tags:
        - 'Table Reservation'
      operationId: a2dd300e428e057b6184d7fa40eacc66
      parameters:
        -
          name: id
          in: path
          description: 'ID du la reservation liée à la table à modifier'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              required:
                - table_id
                - reservation_id
              properties:
                table_id:
                  type: intgeger
                reservation_id:
                  type: integer
              type: object
              example:
                table_id: 3
                reservation_id: 2
      responses:
        '200':
          description: 'Table réservée modifiée'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
    delete:
      tags:
        - 'Table Reservation'
      operationId: 87c1ce67b99e86a4e7f9ee8bfa6e3f67
      parameters:
        -
          name: id
          in: path
          description: 'ID de la réservation liée à la table'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Table réservée supprimée'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
components:
  schemas:
...
    TableReservation:
      properties:
        id:
          type: integer
          nullable: false
        reservation_id:
          type: integer
          nullable: false
        table_id:
          type: integer
          nullable: false
      type: object
...
  securitySchemes:
    Authorization:
      type: http
      bearerFormat: JWT
      scheme: bearer
...